Move compnents dedicated  to settings to their own folder
