+++ src/renderer/settings/TestSpeakersButton.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 // @ts-ignore
-import chime from '../../static/chime.mp3';
-import { ExtendedAudioElement } from './Voice';
+import chime from '../../../static/chime.mp3';
+import { ExtendedAudioElement } from '../Voice';
 
 interface TestSpeakersProps {
 	speaker: string
